refactor: split environmentController into new admin folder for controller
