+++ EMS/core-bundle/src/Resources/config/controllers.xml
@@ -5,6 +5,26 @@
         https://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
+        <!-- Admin -->
+        <service id="EMS\CoreBundle\Controller\Admin\EnvironmentController" public="true">
+            <argument type="service" id="emsco.logger" />
+            <argument type="service" id="ems.service.environment" />
+            <argument type="service" id="ems.service.contenttype" />
+            <argument type="service" id="ems.service.index" />
+            <argument type="service" id="ems.service.mapping" />
+            <argument type="service" id="ems.service.alias" />
+            <argument type="service" id="ems.service.job" />
+            <argument type="service" id="EMS\CoreBundle\Repository\EnvironmentRepository" />
+            <argument type="service" id="ems.repository.field_type" />
+            <argument type="service" id="EMS\CoreBundle\Repository\ContentTypeRepository" />
+            <argument>%ems_core.instance_id%</argument>
+            <argument>%ems_core.circles_object%</argument>
+            <argument>%ems_core.template_namespace%</argument>
+            <argument type="service" id="emsco.data_table.factory" />
+            <call method="setContainer"/>
+            <tag name="container.service_subscriber"/>
+        </service>
+
         <!-- Api\Admin -->
         <service id="EMS\CoreBundle\Controller\Api\Admin\EntitiesController">
             <argument type="service" id="emsco.helper.entities" />
@@ -158,17 +178,8 @@
             <argument type="service" id="ems.service.contenttype" />
             <argument type="service" id="security.authorization_checker" />
             <argument type="service" id="ems.service.publish" />
-            <argument type="service" id="ems.service.index" />
-            <argument type="service" id="ems.service.mapping" />
-            <argument type="service" id="ems.service.alias" />
-            <argument type="service" id="ems.service.job" />
             <argument type="service" id="EMS\CoreBundle\Repository\RevisionRepository" />
-            <argument type="service" id="EMS\CoreBundle\Repository\EnvironmentRepository" />
-            <argument type="service" id="ems.repository.field_type" />
-            <argument type="service" id="EMS\CoreBundle\Repository\ContentTypeRepository" />
             <argument>%ems_core.paging_size%</argument>
-            <argument>%ems_core.instance_id%</argument>
-            <argument>%ems_core.circles_object%</argument>
             <argument>%ems_core.template_namespace%</argument>
             <argument type="service" id="emsco.data_table.factory" />
             <call method="setContainer"/>
