+++ EMS/core-bundle/src/Resources/config/routing/admin/environment.xml
@@ -5,30 +5,30 @@
         xsi:schemaLocation="http://symfony.com/schema/routing https://symfony.com/schema/routing/routing-1.0.xsd">
 
     <route id="environment.attach" path="/attach/{name}"
-           controller="EMS\CoreBundle\Controller\ContentManagement\EnvironmentController::attachAction"
+           controller="EMS\CoreBundle\Controller\Admin\EnvironmentController::attachAction"
            methods="POST"/>
     <route id="environment.remove.alias" path="/remove/alias/{name}"
-           controller="EMS\CoreBundle\Controller\ContentManagement\EnvironmentController::removeAliasAction"
+           controller="EMS\CoreBundle\Controller\Admin\EnvironmentController::removeAliasAction"
            methods="POST"/>
     <route id="environment.remove" path="/remove/{id}"
-           controller="EMS\CoreBundle\Controller\ContentManagement\EnvironmentController::removeAction"
+           controller="EMS\CoreBundle\Controller\Admin\EnvironmentController::removeAction"
            methods="POST"/>
     <route id="environment.add" path="/add"
-           controller="EMS\CoreBundle\Controller\ContentManagement\EnvironmentController::addAction"
+           controller="EMS\CoreBundle\Controller\Admin\EnvironmentController::addAction"
            methods="GET|POST"/>
     <route id="environment.edit" path="/edit/{id}"
-           controller="EMS\CoreBundle\Controller\ContentManagement\EnvironmentController::editAction"
+           controller="EMS\CoreBundle\Controller\Admin\EnvironmentController::editAction"
            methods="GET|POST"/>
     <route id="environment.view" path="/{id}"
-           controller="EMS\CoreBundle\Controller\ContentManagement\EnvironmentController::viewAction"
+           controller="EMS\CoreBundle\Controller\Admin\EnvironmentController::viewAction"
            methods="GET|POST"/>
     <route id="environment.rebuild" path="/rebuild/{id}"
-           controller="EMS\CoreBundle\Controller\ContentManagement\EnvironmentController::rebuild"
+           controller="EMS\CoreBundle\Controller\Admin\EnvironmentController::rebuild"
            methods="GET|POST"/>
     <route id="environment.index" path="/"
-           controller="EMS\CoreBundle\Controller\ContentManagement\EnvironmentController::indexAction"
+           controller="EMS\CoreBundle\Controller\Admin\EnvironmentController::indexAction"
            methods="GET|POST"/>
     <route id="ems_environment_index" path="/"
-           controller="EMS\CoreBundle\Controller\ContentManagement\EnvironmentController::indexAction"
+           controller="EMS\CoreBundle\Controller\Admin\EnvironmentController::indexAction"
            methods="GET|POST"/>
 </routes>
