+++ EMS/core-bundle/src/Resources/config/routing/all.xml
@@ -52,8 +52,8 @@
             prefix="/profile"/>
     <import resource="@EMSCoreBundle/Resources/config/routing/log.xml"
             prefix="/admin/log"/>
-    <import resource="@EMSCoreBundle/Resources/config/routing/environment.xml"
-            prefix="/environment"/>
+    <import resource="@EMSCoreBundle/Resources/config/routing/admin/environment.xml"
+            prefix="/admin/environment"/>
     <import resource="@EMSCoreBundle/Resources/config/routing/publisher.xml"
             prefix="/publisher"/>
     <import resource="@EMSCoreBundle/Resources/config/routing/publish.xml"
