+++ elasticms-admin/config/packages/security.yaml
@@ -85,7 +85,6 @@ security:
         - { path: ^/api, roles: ROLE_API }
         - { path: ^/api/admin, roles: ROLE_ADMIN }
         - { path: ^/admin, roles: ROLE_ADMIN }
-        - { path: ^/environment, roles: ROLE_ADMIN }
         - { path: ^/channel-admin, roles: ROLE_ADMIN }
         - { path: ^/user, roles: ROLE_USER_MANAGEMENT }
         - { path: ^/elasticsearch/delete-search, roles: ROLE_USER }
